Remove & deprecate the Allowlist feature (board, CLI, RPC, service)
The Allowlist feature only ever viewed/aggregated Claude Code permission
patterns from `.claude` settings and promoted a selection into the global
`~/.claude/settings.json`. With most harnesses now running in auto-permission
mode nobody used it, so remove it entirely (CROW-1182).

Removed:
- Dedicated modules: AllowListService, AllowlistRPCSupport, AllowEntry,
  AllowlistCommands, and their tests.
- CLI verbs list-allowlist / promote-allowlist / refresh-allowlist, their RPC
  methods, the `.allowlist` dispatch lane, and the lane-policy entries.
- The sidebar "Allowlist" board, its `#/allowlist` route, and all allowlist
  render/state/RPC/CSS in the web client.
- AppState.allowListSessionID (the …002 virtual-tab UUID), allowEntries /
  isLoadingAllowList, and the DaemonStateSnapshot field.
- The ParityLedger rows so CLI/web parity stays green.

Wire-contract change: an old client calling list-/promote-/refresh-allowlist
now gets unknown-method. Docs/skills/CLAUDE.md/CHANGELOG updated; docs/cli.md
regenerated.

Untouched (different concepts): the MCP tool-catalog allowlist
(MCPToolCatalog, CROW-1004 security boundary), users' actual
`.claude/settings.json` permission rules, and the Settings → Automation
permission auto/allow mode toggles.


Crow-Session: 98A174FF-8D4C-491D-919F-D8FDC0FDE030
