+++ CHANGELOG.md
@@ -29,6 +29,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CROW-980 — The session switcher's default binding is **`esc+tab`** (tap Esc, then Tab), not `shift+tab`. Shift+Tab is already Claude Code's permission-mode cycle, and `switcher.captureInTerminal` defaults to **true**, so the old default ate that keystroke inside every focused terminal. Esc is not a modifier, so this is not a string swap: a leading `esc` is now parsed as a **prefix key**, and `parseSwitcherBinding` returns it separately from `shift/ctrl/alt/meta`. Without that, `esc+tab` would have parsed identically to bare `tab` — every unmodified Tab opening the overlay and, in the terminal, Tab-completion going dead. The prefix arms for 1.5s and is broken by any intervening keystroke (a bare modifier excepted, since a resting Shift shouldn't cancel the sequence), and **Esc itself is never swallowed** — it has to stay instant for interrupting an agent, so it reaches the terminal on the way through and the overlay opens on the Tab that follows. A prefix chord holds nothing down, so it can't commit on release the way `shift+tab` does: the overlay stays open, Tab/←→ cycle, **Enter** (or a click) switches, Esc cancels — which is the ticket's reason for choosing this chord. The hint line under the strip states whichever gesture the configured chord actually uses. Existing installs are untouched: the new default reaches only a config with no `switcher.binding`, and `crow ui set --switcher-binding shift+tab` restores the old one. **Superseded within this same unreleased block by CROW-1002**, which moves the default again (to `cmd+/`) and closes the two holes this entry's last sentence describes: leaving existing installs untouched is precisely what left them eating Shift+Tab, and restoring the old binding is no longer possible — it is reserved.
 
+### Removed
+
+- CROW-1182 — **The Allowlist feature is gone.** The sidebar "Allowlist" board and its `#/allowlist` route, the `crow list-allowlist` / `promote-allowlist` / `refresh-allowlist` CLI verbs, their RPC methods and the dedicated `.allowlist` dispatch lane, and the backing `AllowListService` / `AllowlistRPCSupport` / `AllowEntry` types are all removed. It existed only to view and aggregate Claude Code permission-allowlist patterns from `.claude` settings and promote a selection into the global `~/.claude/settings.json`; with most harnesses now running in auto-permission mode, nobody used it. This is a **wire-contract change**: an old CLI or web client that calls `list-allowlist` / `promote-allowlist` / `refresh-allowlist` now gets an unknown-method error. What it never touched, it still doesn't — users' actual `.claude/settings.json` permission rules are unchanged (Crow only ever read them), the **MCP tool-catalog allowlist** (`MCPToolCatalog`, the CROW-1004 security boundary) is a different concept and is untouched, and the Settings → Automation permission auto/allow **mode** toggles are a separate feature and are unaffected. The `AppState.allowListSessionID` virtual-tab UUID (`…002`) is retired.
+
 ### Fixed
 
 - CROW-1167 — Workspace bootstrap no longer hardcodes the ticket chip as a bare **Issue**. `setup.sh` (and the bundled template) now labels GitHub/GitLab tickets `Issue #<number>` when `--ticket-number` is numeric, matching PR chips (`PR #$N`) and orphan recovery; Jira (no standalone number) still gets `Issue`. The header chip was falling through to this stored link label whenever `ticket_badge` was missing (see #1166).
