+++ CLAUDE.md
@@ -81,18 +81,6 @@ crow transition-ticket --session <uuid> --to inProgress|inReview|done   → move
 crow resync-jira                                                        → re-sync every Jira ticket's status from its Crow session state
 ```
 
-### Allowlist
-
-Claude Code permission patterns, aggregated from global `~/.claude/settings.json` + each worktree's `.claude/settings.local.json`.
-
-```
-crow list-allowlist                                → {"entries":[{pattern,is_global,worktree_session_names}],"loading":bool}; reads the last scan
-crow refresh-allowlist                             → re-scan both sources from disk; {"ok":true}
-crow promote-allowlist --pattern 'Bash(npm test:*)' [--pattern ...]   → {"ok":true,"added":[...],"already_global":[...],"global_settings_path":"..."}
-```
-
-Repeat `--pattern` for more than one, and **quote them** — `(`, `)`, `*` are shell metacharacters. A failed write is a non-zero exit, never `ok:true`; a malformed global settings file is refused rather than overwritten.
-
 ### Settings Commands
 ```
 crow telemetry get                                                      → {"telemetry":{"enabled":…,"port":…,"retention_days":…}}
