+++ Packages/CrowCLI/Sources/CrowCLILib/CrowCommand.swift
@@ -76,9 +76,6 @@ public struct CrowCommand: ParsableCommand {
             EditLink.self,
             TransitionTicket.self,
             ResyncJira.self,
-            ListAllowlist.self,
-            PromoteAllowlist.self,
-            RefreshAllowlist.self,
             GetScorecard.self,
             RebuildScorecard.self,
             GetState.self,
